Separate superstripe regtests
